Move iframe controller tests
